server/ocr/requirements.txt
transformers>=5.0.0
torch>=2.0.0
Pillow
pypdfium2
